interface/pyproject.toml
[tool.poetry]
name = "interface"
version = "0.1.0"
description = ""
authors = ["Maksim Gusev <51376126+AlanTheKnight@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
redis = "^5.0.1"
toml = "^0.10.2"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.11.0"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
